AG-17884 Refactor scene sampler into live + serialized-tree paths
Relocate the SerializedNodeState type family to ag-charts-core (type-only, re-exported
from scene/node.ts) and extract the pure geometry helpers, a generic node-accessor walk,
sampleSerializedRoots and sceneSampleToJSON into _ag-charts-test, so unit and E2E scene
capture share one keying/geometry implementation. The live sampler keeps its per-frame
key stability via the shared walk's WeakMap. Unit snapshot output is unchanged.
