+++ packages/ag-charts-community/src/scene/node.ts
@@ -1,5 +1,8 @@
 import { DeclaredSceneChangeDetection, Logger, createId, createSvgElement, objectsEqual } from 'ag-charts-core';
-import type { AgDrawingMode } from 'ag-charts-types';
+// The serialised-state type family is defined in ag-charts-core so the Node-portable scene sampler
+// (in _ag-charts-test, used by both unit and e2e capture) can type against it without depending on
+// this package. Re-exported here so scene shape files keep importing it from `./node`.
+import type { SerializedNodeProps, SerializedNodeState } from 'ag-charts-core';
 
 import { BBox } from './bbox';
 import type { ImageLoader } from './image/imageLoader';
@@ -44,78 +47,16 @@ export interface NodeOptions {
 
 export type NodeWithOpacity<D> = Node<D> & { opacity: number };
 
-/**
- * Properties every node contributes to its serialised state (see {@link Node.serialize}). The
- * transform properties are contributed by the Translatable/Scalable/Rotatable mixins when applied,
- * so they are optional on every node kind.
- */
-export interface SerializedNodeProps {
-    visible: boolean;
-    translationX?: number;
-    translationY?: number;
-    scalingX?: number;
-    scalingY?: number;
-    rotation?: number;
-}
-
-export interface SerializedGroupProps extends SerializedNodeProps {
-    opacity: number;
-}
-
-export interface SerializedShapeProps extends SerializedNodeProps {
-    opacity: number;
-    drawingMode: AgDrawingMode;
-    hasFill: boolean;
-    hasStroke: boolean;
-}
-
-export interface SerializedPathProps extends SerializedShapeProps {
-    x: number;
-    y: number;
-    width: number;
-    height: number;
-    clip: boolean;
-    clipX: number;
-    clipY: number;
-}
-
-export interface SerializedSectorProps extends SerializedPathProps {
-    startAngle: number;
-    endAngle: number;
-    innerRadius: number;
-    outerRadius: number;
-}
-
-export interface SerializedLineProps extends SerializedShapeProps {
-    x1: number;
-    y1: number;
-    x2: number;
-    y2: number;
-}
-
-export interface SerializedTextProps extends SerializedShapeProps {
-    x: number;
-    y: number;
-    text?: string;
-}
-
-/**
- * Plain-data snapshot of a node's rendered state (see {@link Node.serialize}), discriminated by node
- * kind independently of subclassing (e.g. a specialised bar shape still reads as `'rect'`); `svgPath`
- * carries the drawn path commands in SVG form for path-painting nodes. This union is closed: a new
- * node kind must add its variant here, which forces every exhaustive consumer to decide how to
- * handle it.
- */
-export type SerializedNodeState =
-    | { type: 'node'; props: SerializedNodeProps }
-    | { type: 'group'; props: SerializedGroupProps }
-    | { type: 'path'; props: SerializedPathProps; svgPath?: string }
-    | { type: 'marker'; props: SerializedPathProps; svgPath?: string }
-    | { type: 'rect'; props: SerializedPathProps; svgPath?: string }
-    | { type: 'sector'; props: SerializedSectorProps; svgPath?: string }
-    | { type: 'line'; props: SerializedLineProps }
-    | { type: 'range'; props: SerializedLineProps }
-    | { type: 'text'; props: SerializedTextProps };
+export type {
+    SerializedNodeProps,
+    SerializedGroupProps,
+    SerializedShapeProps,
+    SerializedPathProps,
+    SerializedSectorProps,
+    SerializedLineProps,
+    SerializedTextProps,
+    SerializedNodeState,
+} from 'ag-charts-core';
 
 export type ChildNodeCounts = {
     groups: number;
